unlimited range checking added
git-svn-id: https://svn.shadanakar.org/onPHP/trunk@5533 db88c99c-74ec-0310-8720-8416fdbde762

+++ core/OSQL/TimeIntervalsGenerator.class.php
@@ -92,6 +92,11 @@
 					'define time range and interval units first'
 				);
 			
+			if (!$this->range->getStart() || !$this->range->getEnd())
+				throw new WrongArgumentException(
+					'cannot operate with unlimited range'
+				);
+			
 			$firstIntervalStart =
 				$this->interval->truncate(
 					$this->range->getStart(), !$this->overlapped
